Monitora-backend/requirements.txt
--find-links wheelhouse
alembic==0.8.3
aniso8601==1.1.0
Flask==0.10.1
Flask-Migrate==1.6.0
Flask-RESTful==0.3.4
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.3
MarkupSafe==0.23
marshmallow==2.2.0
marshmallow-jsonapi==0.3.0
python-dateutil==2.5.0
python-editor==0.4
pytz==2015.7
six==1.10.0
SQLAlchemy==1.0.9
Werkzeug==0.15.6
flask-cors==3.0.4
pyodbc==4.0.23
numpy==1.17.0
pandas==0.24.2
ufal.morphodita==1.9.2.3
unidecode==1.1.1


